修改message结构
