+++ processor/handler.go
@@ -135,20 +135,6 @@ func (h *Handler) handleMessage(msg *message.Message, handlerFn HandlerFunc) {
 	//msg.Ack()
 }
 
-// 关于转发handler处理返回的[]*Message，有两种方法
-// 1.将返回的producedMessages循环发布(forwardMessage)
-// 2.以中间件的形式发布(Forward.Middleware)
-// 两者在实现的复杂度上差不多，但在维护的复杂度上，我认为只使用中间件这一种形式更利于阅读和扩展
-//
-//func (h *Handler) forwardMessage(msgs []*message.Message) {
-//	for _, forwarder := range h.forwarders {
-//		err := forwarder.publisher.Publish(forwarder.topic, msgs...)
-//		if err != nil {
-//			return
-//		}
-//	}
-//}
-
 func (h *Handler) handleClose(ctx context.Context) {
 	select {
 	case <-ctx.Done():
